Added a Python script to generate Verilog of Arbiter
This arbiter can be used to replace the arbiter in the multi-layer AHB Bus Matrix offered by ARM System Design Kit. This arbiter appears with the same interface/portlist aiming to easy replacement/integration. However, an arbitration strategy of combinatorial logic is implemented with the advantage of saving one cycle. Namely the arbitration results addr_in_port and no_port are available "instantly" within the same cycle the requests comes, rather than a cycle delay in the ARM original one. The downside is longer combinatorial path to balance, along with a bit overheads.